backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
anthropic==0.52.0
openai==1.82.0
google-cloud-firestore==2.20.1
pydantic==2.11.1
pydantic-settings==2.9.1
sse-starlette==2.2.1
httpx==0.28.1
python-multipart==0.0.20
gunicorn==23.0.0
